Add two additional options from @the-moog, see issue #17
"""
It seems the latest FT_PROG on the FT2232H uses a different CRC to this program.
ftx_prog always reports the CRC wrong.
Currently ftx_prog quits if the CRC is wrong, preventing you from making any changes.

Also if you change the PID, then FT_PROG (wierdly) sometimes fails to see devices - methinks if the CRC is wrong, which it usually is. So there is no way to re-program the device. ftx_prog to the rescue....

Hence ftx_prog is a useful way to restore a bricked device.

The attached diff solves these issues and adds two functions:
--erase-eeprom
--ignore-crc-error
"""